// src/interfaces/anamneses.ts
export interface AnameseProps {
  diabetes: boolean;
  cardiaco: boolean;
  uso_medicacao: boolean;
  exame_period: boolean;
  exame_period_ultim: Date;
  alergia_med: boolean;
  alergia_med_nome: string;
  funcionamento_intestino: string;
  hipertensao: boolean;
  tabagista: boolean;
  etilista: boolean;
  covid: boolean;
  exercio_fisico: boolean;
  oncologico: boolean;
}